backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==3.2.2
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
cryptography==41.0.7
ecdsa==0.19.1
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.104.1
h11==0.16.0
httptools==0.6.4
idna==3.11
numpy==1.24.4
openpyxl==3.1.5
pandas==1.5.3
passlib==1.7.4
pillow==10.4.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
reportlab==3.6.13
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-extensions==4.13.2
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==0.24.0
websockets==12.0
